moved auth out of main layout sections
