+++ client/src/components/application/amendment/AmendmentWorkflow.tsx
@@ -3,7 +3,7 @@ import { ApplicationStatusBadge } from "components/badge/ApplicationStatusBadge"
 import { PhaseSelector, WorkflowApplication } from "components/application";
 import type { Amendment, DemonstrationTypeAssignment } from "demos-server";
 import { gql, useQuery } from "@apollo/client";
-import { Loading } from "components/loading/Loading";
+import { Loading } from "components/loading";
 import { WORKFLOW_PHASE_FIELDS, WORKFLOW_DOCUMENT_FIELDS } from "fragments";
 import { Demonstration } from "pages/DemonstrationsPage";
