+++ client/src/pages/deliverables/DeliverableDetailHeader.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { gql, useQuery } from "@apollo/client";
-import { Loading } from "components/loading/Loading";
+import { Loading } from "components/loading";
 import { DemonstrationDetailHeader } from "pages/DemonstrationDetail/DemonstrationDetailHeader";
 
 export const DELIVERABLE_DETAIL_HEADER_QUERY_NAME = "DeliverableDetailHeader";
